Develop

+++ flaskr/update.py
@@ -1,10 +1,10 @@
 import os
 from pathlib import Path
-from flask import Blueprint, request, make_response
+from flask import Blueprint, request, make_response, g, abort
 
 from os.path import isfile, join
 
-from flaskr.auth import UserResult, login_required
+from flaskr.auth import UserData, login_required
 from flaskr.db import get_db
 from .classes.preProcessClass import PreProcess
 
@@ -21,6 +21,10 @@ bp = Blueprint("update", __name__, url_prefix="/update")
 def delete_file():
     id = request.args.get('id')
     name = request.args.get('name')
+
+    UserData.delete_preprocess_file(id, name)
+    UserData.delete_result(id, name)
+    # UserData.delete_model(id, name)
 
     f_path = USER_PATH / id / name
 
@@ -53,11 +57,21 @@ def update_given_name():
 @login_required
 def delete_user_account():
     id = request.args.get('id')
-    UserResult.remove_user(id)
+    UserData.remove_user(id)
     dir_path = USER_PATH / str(id)
     delete_folder(dir_path)
+
+    delete_user_file(id)
+
     return '1'
 
+def delete_user_file(user_id):
+    files = UserData.get_user_file(user_id)
+
+    for f in files:
+        path = Path(f['path'])
+        if os.path.exists( path ):
+            os.remove(path)
 
 def delete_folder(dir_path):
     try:
@@ -74,7 +88,7 @@ def download_df():
     name = request.args.get('name')
     isTmp = request.args.get('is_tmp')
 
-    if isTmp == 1:
+    if int(isTmp) == 1:
         path = USER_PATH / str(id) / "tmp" / name
         df = PreProcess.getDF(path)
     else:
@@ -108,10 +122,19 @@ def update_user_tour():
 
     return str(want_tour)
 
+def is_not_admin(user):
+    if user['is_admin'] == 0:
+        return True
+    else:
+        return False
 
 @bp.route("/user/admin/", methods=["GET"])
 @login_required
 def update_user_admin():
+    #Check whether admin
+    if is_not_admin(g.user):
+        return abort('401')
+
     id = request.args.get('id')
     is_admin = request.args.get('is_admin')
 
@@ -129,10 +152,22 @@ def update_user_admin():
 @login_required
 def delete_user_files():
 
+    if is_not_admin(g.user):
+        return abort('401')
+
     id = request.args.get('id')
     delete_user_all_files(id)
 
     return str(1)
+
+@bp.route("/delete/tmp/", methods=["GET"])
+@login_required
+def delete_user_tmp_files():
+    id = request.args.get('id')
+    dir_path = USER_PATH / str(id) / "tmp"
+    delete_files_in_dir(dir_path, True)
+
+    return '1'
 
 def delete_user_all_files(id):
     dir_path = USER_PATH / str(id)
@@ -140,8 +175,11 @@ def delete_user_all_files(id):
     dir_path = USER_PATH / str(id) / "tmp"
     delete_files_in_dir(dir_path)
 
-def delete_files_in_dir(path):
+def delete_files_in_dir(path, modal=False):
     for f in os.listdir(path):
+        if modal:
+            if f == '_model.pkl':
+                continue
         file = join(path, f)
         if isfile(file):
             os.remove(file)
@@ -156,7 +194,7 @@ def infrequent_files_delete():
     for id in id_array:
         delete_user_all_files(id)
 
-    UserResult.infrequent_users(ids)
+    UserData.infrequent_users(ids)
 
     return "1"
